use `getIdentStr()`
Spent half an hour trying to figure out why the implementation does not
work properly, only to realize `getStr()` raises 'recoverable'
'error' (`ERecoverableError`), which (IMO) are a mutually exclusive
words, but what do I know. Anyway, because `getStr()` had been defined
for strings, it naturally failed to work with regular identifiers, and
promptly failed when I tried to figure out the name of the field. But
because error was 'recoverable' it bounced off some part of the
execution trace, and got swallowed somewhere.

This is a good example of how simple misuse of the code, paired with
'recoverable' error approach can cause issues that are almost impossible
to debug cleanly, since there are little to no observable side effects.
All failure handling in the compiler must be explicit, and 80-or-so
calls to the `globalError` should be eradicated in the future, because
each of them might silently throw you in completely different part of
the compiler code.